Leverage structural typing using netural serializer types
