+++ src/extension/ai/converters.ts
@@ -7,10 +7,10 @@ import * as parser_pb from '@buf/stateful_runme.bufbuild_es/runme/parser/v1/pars
 
 import { ServerLifecycleIdentity, getServerConfigurationValue } from '../../utils/configuration'
 import { Serializer } from '../../types'
+import { ConnectSerializer } from '../serializer'
 import * as parserTypes from '../grpc/parser/tcp/types'
 import * as serializer from '../serializer'
 import { Kernel } from '../kernel'
-import { GrpcSerializer } from '../serializer'
 
 // Converter provides converstion routines from vscode data types to protocol buffer types.
 // It is a class because in order to handle the conversion we need to keep track of the kernel
@@ -29,7 +29,7 @@ export class Converter {
     let notebookDataWithExec = new vscode.NotebookData(cellDataWithExec)
     // marshalNotebook returns a protocol buffer using the ts client library from buf we need to
     // convert it to es
-    return GrpcSerializer.marshalNotebook(notebookDataWithExec)
+    return ConnectSerializer.marshalNotebook(notebookDataWithExec)
   }
 }
 
@@ -80,7 +80,7 @@ export function cellProtosToCellData(cells: parserTypes.Cell[]): vscode.Notebook
     'lifecycleIdentity',
     parserTypes.RunmeIdentity.ALL,
   )
-  let newCellData = GrpcSerializer.revive(notebook, identity)
+  let newCellData = ConnectSerializer.revive(notebook, Number(identity))
   return newCellData
 }
